タスクコントローラーのリファクタリングを行い、エラーハンドリングとレスポンス処理を改善しました。また、UUIDフィールドのデフォルト値を修正し、子供タスクの追加処理を更新しました。新たにレスポンスハンドラーを追加し、ユーザーコントローラーでも同様の改善を行いました。
